Move test helpers to a shared scope `TestHelper`
